+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ### NEXT
 
+### 3.24.1
+
 - Worker: Don't check `ignoredSubchannel` in piped `DataConsumers` ([PR #1879](https://github.com/versatica/mediasoup/pull/1879)).
 
 ### 3.24.0